--- apps/kittynode-gui/kittynode-frontend/src/components/index.ts
@@ -1,2 +1,0 @@
-export { default as Kaomoji } from "./kaomoji.svelte";
-export { default as Welcome } from "./welcome.svelte";
