--- apps/kittynode-gui/kittynode-frontend/src/components/kaomoji.svelte
@@ -1,7 +1,0 @@
-<div>
-  &nbsp;∧,,,∧&nbsp;~ ┏━━━━━━━━┓
-  <br>
-  (  ̳• · • ̳)&nbsp;~ ♡&nbsp;&nbsp;You're purrfect&nbsp;&nbsp;♡
-  <br>
-  /&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;づ&nbsp;~ ┗━━━━━━━━┛
-</div>
