+++ apps/kittynode-gui/kittynode-frontend/src/routes/+layout.svelte
@@ -1,58 +1,23 @@
 <script lang="ts">
 import "@picocss/pico/css/pico.purple.min.css";
-import { page } from "$app/stores";
 import { onMount } from "svelte";
 import { win } from "../stores/window.svelte.ts";
-import { invoke } from "@tauri-apps/api/core";
-import { Welcome } from "../components";
+import { initializedStore } from "../stores/initialized.svelte";
+import Welcome from "./Welcome.svelte";
+import Header from "./Header.svelte";
 
 const { children } = $props();
-let isInitialized = $state(false);
-
-async function getIsInitialized() {
-  try {
-    isInitialized = await invoke("is_initialized");
-  } catch (error) {
-    alert("Failed to check if Kittynode is initialized.");
-    console.error(error);
-  }
-}
 
 onMount(async () => {
-  await getIsInitialized();
-  win.show();
+  console.log("initializedStore.initialized", initializedStore.initialized);
+  await win.show();
 });
 </script>
 
-{#if !isInitialized}
+{#if !initializedStore.initialized}
 <Welcome />
 {:else}
-<header class="container">
-  <nav>
-    <ul>
-      <li>
-        <strong>
-          Kittynode
-        </strong>
-      </li>
-    </ul>
-    <ul>
-      <li>
-        <a href="/" aria-current={$page.url.pathname === "/"}>Home</a>
-      </li>
-      <li>
-        <a href="/packages" aria-current={$page.url.pathname === "/packages"}>Packages</a>
-      </li>
-      <li>
-        <a href="/system-info" aria-current={$page.url.pathname === "/system-info"}>System Info</a>
-      </li>
-      <li>
-        <a href="/settings" aria-current={$page.url.pathname === "/settings"}>Settings</a>
-      </li>
-    </ul>
-  </nav>
-</header>
-
+<Header />
 <main class="container" style="margin-top: 1rem;">
   {@render children()}
 </main>
