+++ apps/kittynode-gui/kittynode-frontend/src/routes/settings/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 import { invoke } from "@tauri-apps/api/core";
 import { message } from "@tauri-apps/plugin-dialog";
+import { initializedStore } from "../../stores/initialized.svelte";
 
 async function connectMobile() {
   await message("Coming soon.");
@@ -9,7 +10,8 @@ async function connectMobile() {
 async function deleteKittynode() {
   try {
     await invoke("delete_kittynode");
-    window.location.reload();
+    await initializedStore.uninitialize();
+    message("Kittynode data has been deleted successfully.");
   } catch (error) {
     alert("Failed to delete Kittynode.");
     console.error(error);
