+++ apps/kittynode-gui/kittynode-frontend/src/stores/initialized.svelte.ts
@@ -1,0 +1,20 @@
+import { invoke } from "@tauri-apps/api/core";
+
+let initialized = $state(false);
+
+export const initializedStore = {
+  get initialized() {
+    return initialized;
+  },
+  async initialize() {
+    if (!initialized) {
+      await invoke("init_kittynode");
+      initialized = true;
+    }
+  },
+  async uninitialize() {
+    if (initialized) {
+      initialized = false;
+    }
+  },
+};
