+++ apps/kittynode-gui/kittynode-tauri/tauri.conf.json
@@ -12,7 +12,8 @@
     "windows": [
       {
         "title": "Kittynode",
-        "visible": false
+        "visible": false,
+        "transparent": true
       }
     ],
     "security": {
